DEV-31835 Allow override of Maven URL from where artifacts are fetched
Currently, the remote search URL of Maven central returns a URL to
repo2.maven.org and this domain has an invalid SSL certificate. This
causes the Chef recipe to error out since the underlying Chef resource
properly enforces hostname validation on the resulting SSL certificate.

This change removes the search endpoint in favor of directly downloading
from the repo1.maven.org domain and allowing this to be customized as an
attribute.